fix: move stream_chat_completion_chunks out of helpers and into cli
Doesn't belong in library. This is for cli.
