+++ mellea/helpers/openai_compatible_helpers.py
@@ -244,108 +244,3 @@ def build_completion_usage(output: Any) -> CompletionUsage | None:
         completion_tokens=completion_tokens,
         total_tokens=total_tokens,
     )
-
-
-async def stream_chat_completion_chunks(
-    output: Any,
-    completion_id: str,
-    model: str,
-    created: int,
-    stream_options: dict[str, Any] | None = None,
-    system_fingerprint: str | None = None,
-) -> AsyncGenerator[str, None]:
-    """Generate OpenAI-compatible SSE chat completion chunks from a model output.
-
-    Args:
-        output: The model output object to stream.
-        completion_id: Unique identifier for this completion.
-        model: Model name to include in chunks.
-        created: Unix timestamp of when the completion was created.
-        stream_options: OpenAI-compatible streaming options. Currently supports
-            ``include_usage`` (bool) to control whether usage stats are included
-            in the final chunk. Defaults to including usage when available.
-        system_fingerprint: Backend configuration fingerprint to include in chunks.
-            Defaults to ``None``.
-
-    Yields:
-        Server-sent event payload strings representing OpenAI-compatible chat
-        completion chunks, including the terminating ``[DONE]`` event.
-    """
-    from cli.serve.models import (
-        ChatCompletionChunk,
-        ChatCompletionChunkChoice,
-        ChatCompletionChunkDelta,
-        OpenAIError,
-        OpenAIErrorResponse,
-    )
-
-    try:
-        initial_chunk = ChatCompletionChunk(
-            id=completion_id,
-            model=model,
-            created=created,
-            choices=[
-                ChatCompletionChunkChoice(
-                    index=0,
-                    delta=ChatCompletionChunkDelta(role="assistant", content=""),
-                    finish_reason=None,
-                )
-            ],
-            object="chat.completion.chunk",
-            system_fingerprint=system_fingerprint,
-        )
-        yield f"data: {initial_chunk.model_dump_json()}\n\n"
-
-        previous_length = 0
-        while not output.is_computed():
-            new_content = await output.astream()
-            previous_length += len(new_content)
-
-            if new_content:
-                chunk = ChatCompletionChunk(
-                    id=completion_id,
-                    model=model,
-                    created=created,
-                    choices=[
-                        ChatCompletionChunkChoice(
-                            index=0,
-                            delta=ChatCompletionChunkDelta(content=new_content),
-                            finish_reason=None,
-                        )
-                    ],
-                    object="chat.completion.chunk",
-                    system_fingerprint=system_fingerprint,
-                )
-                yield f"data: {chunk.model_dump_json()}\n\n"
-
-        # Include usage in final chunk if requested via stream_options
-        # Default to True (include usage) for backward compatibility
-        include_usage = True
-        if stream_options is not None:
-            include_usage = stream_options.get("include_usage", True)
-
-        usage = build_completion_usage(output) if include_usage else None
-
-        final_chunk = ChatCompletionChunk(
-            id=completion_id,
-            model=model,
-            created=created,
-            choices=[
-                ChatCompletionChunkChoice(
-                    index=0,
-                    delta=ChatCompletionChunkDelta(content=""),
-                    finish_reason="stop",
-                )
-            ],
-            object="chat.completion.chunk",
-            system_fingerprint=system_fingerprint,
-            usage=usage,
-        )
-        yield f"data: {final_chunk.model_dump_json()}\n\n"
-        yield "data: [DONE]\n\n"
-
-    except Exception as e:
-        error_response = OpenAIErrorResponse(
-            error=OpenAIError(message=f"Streaming error: {e!s}", type="server_error")
-        )
-        yield f"data: {error_response.model_dump_json()}\n\n"
